+++ pages/site_sidebar.php
@@ -48,6 +48,30 @@
     ?>
     <li class="treeview">
       <a href="#">
+        <i class="fa fa-cloud"></i> <span>Clientes DC</span>
+        <span class="pull-right-container">
+          <i class="fa fa-angle-left pull-right"></i>
+        </span>
+      </a>
+      <ul class="treeview-menu">
+        <li><a href="cdc_dashboard.php"><i class="fa fa-pie-chart"></i> Dashboard</a></li>
+        <li><a href="cdc_organismo.php"><i class="fa fa-building"></i> Organismos</a></li>
+        <li><a href="cdc_cliente.php"><i class="fa fa-user"></i> Clientes</a></li>
+        <li class="treeview">
+          <a href="#"><i class="fa fa-gears"></i> Servicios
+            <span class="pull-right-container">
+              <i class="fa fa-angle-left pull-right"></i>
+            </span>
+          </a>
+          <ul class="treeview-menu">
+            <li><a href="sdc_housing.php"><i class="fa fa-home"></i> Housing</a></li>
+            <li><a href="sdc_hosting.php"><i class="fa fa-server"></i> Hosting</a></li>
+          </ul>
+        </li>
+      </ul>
+    </li>    
+    <li class="treeview">
+      <a href="#">
         <i class="fa fa-pie-chart"></i><span>Métricas</span>
         <span class="pull-right-container">
           <i class="fa fa-angle-left pull-right"></i>
@@ -59,8 +83,8 @@
         <li><a href="met_riesgos.php"><i class="fa fa-flash"></i> <span>Riesgos</span></a></li>
         <li><a href="met_controles.php"><i class="fa fa-retweet"></i> <span>Controles</span></a></li>
       </ul>
-    </li>;
- 
+    </li>
+
   </ul>
   <!-- /.sidebar-menu -->
 </section>